+++ components/Footer.js
@@ -1,59 +1,29 @@
 import Link from 'next/link';
 import footerStyles from '../styles/Footer.module.scss';
-import { linksNav, linksSocial } from '../utils/links';
+import { linksNav } from '../utils/links';
 import NewsletterSubscribe from './mailchimp/NewsletterSubscribe';
 import Image from 'next/image';
 
 export default function Footer() {
   return (
     <footer className={footerStyles.footer}>
-      <div className={footerStyles.container}>
-        <div className={footerStyles.columnLeft}>
-          <ul className={footerStyles.footerList}>
+      <NewsletterSubscribe />
+      <section className={footerStyles.footer__section}>
+        <nav className={footerStyles.footer__nav} aria-label="Main">
+          <ul className={footerStyles.footer__navList}>
             {linksNav.map(link => (
-              <li className={footerStyles.footerItem} key={link.href}>
-                <Link href={link.href}>
-                  <a title={link.text}>{link.text}</a>
-                </Link>
+              <li className={footerStyles.footer__navItem} key={link.text}>
+                <Link href={link.href}>{link.text}</Link>
               </li>
             ))}
           </ul>
-          <Image
-            width={77}
-            height={72}
-            className={footerStyles.logo}
-            src="/images/web-dev-path-logo-small.png"
-            alt="Logo"
-          />
-        </div>
-        <div className={footerStyles.columnRight}>
-          <div className={footerStyles.subscribe}>
-            <h2>Subscribe for more</h2>
-            <p className={footerStyles.text}>
-              Get the answer to the most common questions directly to your email
-            </p>
-            <NewsletterSubscribe />
-            <p>*Unsubscribe anytime</p>
-          </div>
-          <div className={footerStyles.socialMedia}>
-            {linksSocial.map(link => (
-              <Link href={link.href} key={link.text} passHref>
-                <a>
-                  <Image
-                    width={32}
-                    height={32}
-                    src={link.src}
-                    alt={link.text}
-                  />
-                </a>
-              </Link>
-            ))}
-          </div>
-        </div>
-      </div>
-      <div className={footerStyles.copyright}>
-        <p>Web Dev Path {new Date().getFullYear()}. All rights reserved.</p>
-      </div>
+        </nav>
+        <Link href="/">
+          <a className={footerStyles.footer__logo} title="Go to the Homepage">
+            <Image src="/images/svg/logo.svg" height={326} width={326} />
+          </a>
+        </Link>
+      </section>
     </footer>
   );
 }