+++ utils/links.js
@@ -1,4 +1,8 @@
 export const linksNav = [
+  {
+    text: 'Join us',
+    href: 'https://webdevpath.slack.com/join/shared_invite/zt-xqqgwwo5-a09BSVWC9ZrHmS6RaMBzVw#/shared-invite/email',
+  },
   { text: 'About Us', href: '/about-us' },
   { text: 'Blog', href: '/blog' },
   { text: 'Contact Us', href: '/contact-us' },